fix(settings): stop the DHCP pool-size cap wrapping past 65536
be32(end) - be32(start) + 1 ran in uint32, so a range spanning the
whole IPv4 space (0.0.0.0-255.255.255.255) wrapped to 0 and slipped
past the 65536-address cap R4 exists to enforce. Widen to uint64
before subtracting so it can no longer wrap, and add rejection tests
for 65537 addresses and the whole address space (the regression case
for this bug).

+++ internal/settings/validate.go
@@ -144,8 +144,8 @@ func validateDHCP(v store.Settings) error {
 	// The lease table is bounded by the range size, so cap the pool rather
 	// than requiring one /24: SuggestRange can propose a range that spans
 	// two /24s on anything wider than a /23, and that is legitimate.
-	size := be32(end.As4()) - be32(start.As4()) + 1
-	if size > dhcpMaxPoolSize {
+	size := uint64(be32(end.As4())) - uint64(be32(start.As4())) + 1
+	if size > uint64(dhcpMaxPoolSize) {
 		return bad("dhcp.range_end", "the range holds %d addresses, more than the %d-address limit", size, dhcpMaxPoolSize)
 	}
 	if _, err := parseIPv4("dhcp.gateway", v.DHCPGateway); err != nil {
